+++ svd-encoder/src/field.rs
@@ -7,8 +7,8 @@ impl Encode for Field {
 
     fn encode(&self) -> Result<Element, EncodeError> {
         match self {
-            Field::Single(info) => info.encode(),
-            Field::Array(info, array_info) => {
+            Self::Single(info) => info.encode(),
+            Self::Array(info, array_info) => {
                 let mut base = Element::new("field");
                 base.merge(&array_info.encode()?);
                 base.merge(&info.encode()?);