+++ svd-parser/src/lib.rs
@@ -162,6 +162,7 @@ mod fieldinfo;
 mod interrupt;
 mod modifiedwritevalues;
 mod peripheral;
+mod peripheralinfo;
 mod register;
 mod registercluster;
 mod registerinfo;